requirements.txt
asgiref
black==21.8b0
click==8.0.1
colorama==0.4.4
Django==5.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.3.0
pytz==2021.1
regex==2021.8.28
sqlparse
tomli==1.2.1
typing-extensions==3.10.0.2
six==1.16.0
openpyxl==3.0.10
pandas
fpdf2
